refactor(user-invitation): modify frontend to work with updated backend
