+++ helpdesk/hooks.py
@@ -42,7 +42,8 @@ website_route_rules = [
 ]
 
 user_invitation = {
-    "allowed_invite_roles": ["System Manager", "Agent", "Agent Manager"],
+    "only_for": ["Agent Manager", "System Manager"],
+    "allowed_roles": ["Agent", "Agent Manager", "System Manager"],
     "after_accept": "helpdesk.helpdesk.hooks.user_invitation.after_accept",
 }
