+++ helpdesk/patches.txt
@@ -20,5 +20,5 @@ helpdesk.helpdesk.doctype.hd_ticket.patches.first_responded_on
 helpdesk.patches.update_hd_team_users
 helpdesk.patches.default_article_category #v2
 helpdesk.patches.communication_read_perm_agent
-helpdesk.patches.agent_manager_perms
-helpdesk.patches.remove_agents_teams_default_views 
+helpdesk.patches.agent_manager_perms #v2
+helpdesk.patches.remove_agents_teams_default_views
